[#13] Added MarkdownEditor.vue
